requirements.txt
alembic==1.0.3
Click==7.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.14
Werkzeug==0.14.1
alembic==1.0.3
Click==7.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
mysql-connector-python-rf==2.2.2
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.14
Werkzeug==0.14.1
